Allow additional EnvVar config in container_config tags
Summary:
Use the new model building utility to give the full range of V1EnvVar functionality in tag config

Test Plan: BK
