+++ integration_tests/python_modules/dagster-k8s-test-infra/dagster_k8s_test_infra/integration_utils.py
@@ -48,20 +48,3 @@ def within_docker():
         or os.path.isfile(cgroup_path)
         and any("docker" in line for line in open(cgroup_path))
     )
-
-
-def remove_none_recursively(obj):
-    """Remove none values from a dict. This is used here to support comparing provided config vs.
-    config we retrive from kubernetes, which returns all fields, even those which have no value
-    configured.
-    """
-    if isinstance(obj, (list, tuple, set)):
-        return type(obj)(remove_none_recursively(x) for x in obj if x is not None)
-    elif isinstance(obj, dict):
-        return type(obj)(
-            (remove_none_recursively(k), remove_none_recursively(v))
-            for k, v in obj.items()
-            if k is not None and v is not None
-        )
-    else:
-        return obj
